YaraDiskScan/requirements.txt
dissect-target>=3.0.0
yara-python>=4.3.0
